Convert --nice option parsing to new form.
Note that this intentionally discards the checks for a nice
value < 0, which were previously limited to SlurmUser / root.

However - an Admin is allowed by slurmctld to submit with a lower
value, or they can alter the job after submission to change the nice
value further. So get rid of that check here.